Refactored MpHealth Endpoint FAT to use common HealthFileUtils - Co-authored-by-AI: IBM Bob 1.0.3
