+++ dev/io.openliberty.microprofile.health.file.healthcheck.endpoints.internal_fat/fat/src/io/openliberty/microprofile/health/file/healthcheck/endpoints/fat/FATSuite.java
@@ -22,7 +22,7 @@ import org.junit.runners.Suite.SuiteClasses;
 import com.ibm.websphere.simplicity.log.Log;
 
 import componenttest.custom.junit.runner.AlwaysPassesTest;
-import io.openliberty.microprofile.health.file.healthcheck.endpoints.fat.utils.HealthFileUtils;
+import io.openliberty.microprofile.health.internal_fat.shared.HealthFileUtils;
 
 @RunWith(Suite.class)
 @SuiteClasses({
